requirements.txt
selenium==4.1.0
uvicorn==0.17.6
webdriver-manager==3.7.0
mysql-connector-python==8.0.29
fastapi==0.78.0
install-jdk==0.3.0
pytest==7.1.2
